HW-6 TickerItem 바인딩 어댑터 적용
